fix: move applyPassive to dom package for use in text-field
