+++ packages/mdc-dom/events.ts
@@ -1,6 +1,6 @@
 /**
  * @license
- * Copyright 2018 Google Inc.
+ * Copyright 2019 Google Inc.
  *
  * Permission is hereby granted, free of charge, to any person obtaining a copy
  * of this software and associated documentation files (the "Software"), to deal
@@ -21,7 +21,32 @@
  * THE SOFTWARE.
  */
 
-import * as events from './events';
-import * as ponyfill from './ponyfill';
+/**
+ * Stores result from applyPassive to avoid redundant processing to detect
+ * passive event listener support.
+ */
+let supportsPassive_: boolean | undefined;
 
-export {events, ponyfill};
+/**
+ * Determine whether the current browser supports passive event listeners, and
+ * if so, use them.
+ */
+export function applyPassive(globalObj: Window = window, forceRefresh = false):
+    boolean | EventListenerOptions {
+  if (supportsPassive_ === undefined || forceRefresh) {
+    let isSupported = false;
+    try {
+      globalObj.document.addEventListener('test', () => undefined, {
+        get passive() {
+          isSupported = true;
+          return isSupported;
+        },
+      });
+    } catch (e) {
+    } // tslint:disable-line:no-empty cannot throw error due to tests. tslint also disables console.log.
+
+    supportsPassive_ = isSupported;
+  }
+
+  return supportsPassive_ ? {passive: true} as EventListenerOptions : false;
+}
